Implemented additional genetic operator types.
Added SQL functionality for databases creation.

Fixed best solution issue in mutation method.

Modified crossover to be pairwise, to ensure consisted GO performances between runs.

+++ XRDInterfaces/TOPAS.py
@@ -1,15 +1,18 @@
 import re
 import os
 import subprocess
+from subprocess import Popen
 import string
 import time
 from pprint import pprint
 
+import PyCrystGA
 from Config.Config import *
 
 class TOPAS:
-    def __init__(self, topasPath, templateFile, seqFile):
+    def __init__(self, topasPath, templateFile, seqFile, topasVersion):
         self.directory = Config.get('working_directory') + 'XRD/' #@todo this is the problem
+        self.topasVersion = int(topasVersion)
         self.topasPath = topasPath
         self.templateFile = templateFile
         self.seqFile = seqFile
@@ -88,9 +91,12 @@ class TOPAS:
 
 
     def makeInputFile(self, structure):
+        fileOutput = '\s*out\s*-'
         taString = 'Rotate_about_points'
         rotString = 'rotate'
         tranString = 't[a-c]'
+
+        fileRegex = re.compile(fileOutput)
         taRegex = r'(\b' + taString + r'\s*\(\s*\s+)([+-]?\d+\.?\d*)'
         rotRegex = r'(\b' + rotString + r'\s*\s+)([+-]?\d+\.?\d*)'
         tranRegex = r'(\b' + tranString + r'\s*\s+)([+-]?\d+\.?\d*)'
@@ -103,6 +109,7 @@ class TOPAS:
         ta = re.compile(taRegex)
         rot = re.compile(rotRegex)
         tran = re.compile(tranRegex)
+        file = re.compile(fileRegex)
 
         replaceTa = []
         replaceAng = []
@@ -144,27 +151,48 @@ class TOPAS:
                 line = tran.sub(replaceTran[z], line)
                 z += 1
 
+            d = file.search(line)
+
+            if d:
+                line = file.sub('   out ' + '\"' + self.directory + "rwp.txt" + '\"' + " append", line)
+
             newContents.append(line)
 
-        with open((self.getInputFilePath(structure.identifier)), 'w') as out:
+        file1 = self.getInputFilePath(structure.identifier)
+
+        with open(file1, 'w') as out:
             out.writelines(newContents)
 
         return self.getInputFilePath(structure.identifier)
 
-    def makeBatchFile(self, population):
+    def makeBatchFile(self, structureList):
+        #@todo figure out how to run batch file from within python
         print('Batch')
-        myBat = open(self.directory + "TOPASBatch.bat", 'w+')
-        for structure in population:
-            print(str(self.topasPath + " " +self.getInputFilePath(structure.identifier)))
-            myBat.write(self.topasPath + " " +self.getInputFilePath(structure.identifier))
+        batFile = "C:/Topas-6/TOPASBatch.bat"
+        print(batFile)
+        myBat = open(batFile, 'w+')
+
+        myBat.write("CD " + self.topasPath + "\n")
+        for structure in structureList:
+            self.makeInputFile(structure)
+        for structure in structureList:
+            myBat.write("tc " + self.getInputFilePathBatch(structure.identifier) + "\n")
         myBat.close()
         return
+
+    def runBatchFile(self):
+        p = Popen(self.topasPath + "TOPASBatch.bat", cwd=self.topasPath,
+                  stderr=subprocess.STDOUT, stdout=subprocess.DEVNULL)
+        p.communicate()
 
     def getInputFilePath(self, identifier):
         return self.directory + str(identifier) + '.inp'
 
+    def getInputFilePathBatch(self, identifier):
+        return self.directory + str(identifier) + '.inp'
+
     def runInputFile(self, inputFile):
-        subprocess.run(os.path.normpath(self.topasPath) + ' ' + os.path.normpath(
+        subprocess.run(os.path.normpath(self.topasPath + "tc.exe") + ' ' + os.path.normpath(
             inputFile), stderr=subprocess.STDOUT, stdout=subprocess.DEVNULL)
         # subprocess.run(os.path.normpath(self.topasPath) + ' ' + os.path.normpath(
         #     inputFile))
@@ -184,42 +212,62 @@ class TOPAS:
         fOpen.close()
         return newRwp
 
-    def newEvaluate(self, inputFile):
-        self.runInputFile(inputFile)
+    def newEvaluate(self, inputFile, structureList, population):
+
+        if self.topasVersion < 6:
+            self.makeBatchFile(structureList)
+            self.runBatchFile()
+        else:
+            self.runInputFile(inputFile)
+
         fitnessFile = self.directory + "rwp.txt"
         fOpen = open(fitnessFile, "r")
-        rwp = []
-        for line in fOpen:
-            rwp.append((1. /float(line),))
-        # print('zap')
-        # print(rwp)
-        fOpen.close()
-        if os.path.exists(fitnessFile):
-            os.remove(fitnessFile)
-        else:
-            print(fitnessFile)
-        return rwp
+        # rwp = []
 
+        """
+        Extracts the rwp values for the population/sub-population of structure provided
+        """
+        for line in fOpen:
+            population.rwpVals.append((1. /float(line),))
+        fOpen.close()
+
+        # print("list rwp")
+        # print(len(rwp))
+        # print(rwp)
+        """
+        file directory clean up
+        """
+        if self.topasVersion >= 6:
+            if os.path.exists(fitnessFile):
+                os.remove(fitnessFile)
+            else:
+                print(fitnessFile)
+
+        if self.topasVersion < 6:
+            for f in os.listdir(self.directory):
+                os.remove(os.path.join(self.directory, f))
+        # return rwp
+        return
 
 
         # rwp = self.getRwpFromInputFile(inputFile)
         # self.fileRemoval(structure)
 
-    def evaluate(self, structure):
-        start = time.time()
-        inputFile = self.makeInputFile(structure)
-        t1 = time.time()
-        time1 = t1-start
-        # print("time1")
-        # print(time1)
-        self.runInputFile(inputFile)
-        t2 = time.time()
-        time2 = t2-t1
-        # print('time2')
-        # print(time2)
-        rwp = self.getRwpFromInputFile(inputFile)
-        self.fileRemoval(structure)
-        return 1. / float(rwp),
+    # def evaluate(self, structure):
+    #     start = time.time()
+    #     inputFile = self.makeInputFile(structure)
+    #     t1 = time.time()
+    #     time1 = t1-start
+    #     # print("time1")
+    #     # print(time1)
+    #     self.runInputFile(inputFile)
+    #     t2 = time.time()
+    #     time2 = t2-t1
+    #     # print('time2')
+    #     # print(time2)
+    #     rwp = self.getRwpFromInputFile(inputFile)
+    #     self.fileRemoval(structure)
+    #     return 1. / float(rwp),
 
     def fileRemoval(self, structure):
         os.remove(self.directory + str(structure.identifier) + '.inp')
